+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for model import from parquet file metadata.
 - Changelog support for `Info` and `Terms` blocks.
+- Added import for `ProtoBuf`
 
 ### Changed
 - Changelog support for custom extension keys in `Models` and `Fields` blocks.